Use static injection for Responder plugin types
